pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-ctl-financeiro"
version = "0.1.0"
description = "Financeiro"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas",
    "psycopg2-binary",
    "python-dotenv"
]

[tool.setuptools]
packages = ["src"]